requirements.txt
flet
requests
Pillow
beautifulsoup4
reportlab
